refactor: refactor time picker field component to a separate file
